Using parenthesis for method calls is optional
Remove the requirement to always use parenthesis in method calls with arguments.

There are a lot of tradeoffs for using/excluding parenthesis, and it results in a lot of personal preference. By removing this requirement developers must exercise their own judgement in writing clear, readable, maintainable code. Code must always be clear, readable, and maintainable, regardless of whether braces are included or not.

Arguments put forward in favour of removing this (in no particular order):

  * It can make code more like English; thus, making it faster and easier to read for some developers.
  * Following the ruby community conventions is usually a good idea.
  * It can add unnecessary clutter to the code (code junk).

Arguments put forward in favour of keeping it (in no particular order):

  * Without brackets it looks more like English than code. Some developers find it harder to read.
  * It can be ambiguous and clear about what is happening.
  * Code never breaks by adding brackets, the inverse is not true.